WebM file fail to play https://bugs.webkit.org/show_bug.cgi?id=242854 <rdar://97190939>
Reviewed by NOBODY (OOPS!).

If a track is disabled because it contains no data, the AV MediaPlayer will
flag the content at not playbable even though it will play.

We can't properly test this change for the failing file as we rely on the
MediaFormatReader plugin to decode WebM which isn't available on the EWS bots.

* Source/WebCore/platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.h:
* Source/WebCore/platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.mm:
(WebCore::MediaPlayerPrivateAVFoundationObjC::allTracksArePlayable const):
(WebCore::MediaPlayerPrivateAVFoundationObjC::containsDisabledTracks const):
(WebCore::MediaPlayerPrivateAVFoundationObjC::assetStatus const):